Cleanup in UI files
* Remove "list" components similar to what was already done for SkillList in previous PR. The list is now directly on the associated page.
* Clean up imports section. Add section for imported types, group material UI imports together.
