--- src/Bladeburner/ui/GeneralActionList.tsx
@@ -1,19 +1,0 @@
-import React from "react";
-import { GeneralActionElem } from "./GeneralActionElem";
-import { GeneralActions } from "../data/GeneralActions";
-import { Bladeburner } from "../Bladeburner";
-
-interface GeneralActionListProps {
-  bladeburner: Bladeburner;
-}
-
-export function GeneralActionList({ bladeburner }: GeneralActionListProps): React.ReactElement {
-  const actions = Object.values(GeneralActions);
-  return (
-    <>
-      {actions.map((action) => (
-        <GeneralActionElem key={action.name} bladeburner={bladeburner} action={action} />
-      ))}
-    </>
-  );
-}
