+++ src/Bladeburner/ui/KillIcon.tsx
@@ -1,8 +1,6 @@
 import React from "react";
+import { Tooltip, Typography } from "@mui/material";
 import { killIcon } from "../data/Icons";
-
-import Typography from "@mui/material/Typography";
-import Tooltip from "@mui/material/Tooltip";
 
 export function KillIcon(): React.ReactElement {
   return <Tooltip title={<Typography>This action involves retirement</Typography>}>{killIcon}</Tooltip>;
