+++ src/Bladeburner/ui/OperationPage.tsx
@@ -1,13 +1,16 @@
-import * as React from "react";
-import { OperationList } from "./OperationList";
-import { Bladeburner } from "../Bladeburner";
-import Typography from "@mui/material/Typography";
+import type { Bladeburner } from "../Bladeburner";
 
-interface IProps {
+import * as React from "react";
+import { BladeOperationName } from "@enums";
+import { OperationElem } from "./OperationElem";
+import { Typography } from "@mui/material";
+
+interface OperationPageProps {
   bladeburner: Bladeburner;
 }
 
-export function OperationPage(props: IProps): React.ReactElement {
+export function OperationPage({ bladeburner }: OperationPageProps): React.ReactElement {
+  const names = Object.values(BladeOperationName);
   return (
     <>
       <Typography>
@@ -27,7 +30,9 @@ export function OperationPage(props: IProps): React.ReactElement {
         You can unlock higher-level operations by successfully completing them. Higher-level operations are more
         difficult, but grant more rank and experience.
       </Typography>
-      <OperationList bladeburner={props.bladeburner} />
+      {names.map((name) => (
+        <OperationElem key={name} bladeburner={bladeburner} operation={bladeburner.operations[name]} />
+      ))}
     </>
   );
 }
